한번 더 추가 및 변경: realBird을 주로 참조하여 변경 및 추가를 하였다. image upload 는 일단 public 폴더에 하는 것으로 변경하였다. 추후 cloudflare에 업로드하고 live stream도 추가할 예정이다.

+++ pages/api/chat/index.ts
@@ -78,6 +78,9 @@ async function handler(req: NextApiRequest, res: NextApiResponse<ResponseType>) 
     const {
       query: { roomId },
     } = req;
+    if (!roomId) {
+      return res.status(404).end({ error: "request query is not given." });
+    }
     if (roomId) {
       const delChatRoom = await client.chatRoom.deleteMany({
         where: {
